requirements.txt
# AI 论文追踪项目依赖
requests>=2.31.0
beautifulsoup4>=4.12.0
huggingface_hub>=0.23.0
arxiv>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
